+++ src/Runner.Worker/Container/ContainerHooks/ContainerHookInput.cs
@@ -3,7 +3,7 @@ using System.Runtime.Serialization;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace GitHub.Runner.Worker.Container
+namespace GitHub.Runner.Worker.Container.ContainerHooks
 {
     public class ContainerHookInput
     {
