+++ src/Runner.Worker/Container/ContainerInfo.cs
