+++ src/Runner.Worker/ContainerOperationProvider.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.ServiceProcess;
 using System.Threading.Tasks;
 using System.Linq;
 using System.Threading;
@@ -10,8 +9,8 @@ using GitHub.Services.Common;
 using GitHub.Runner.Common;
 using GitHub.Runner.Sdk;
 using GitHub.DistributedTask.Pipelines.ContextData;
-using Microsoft.Win32;
 using GitHub.DistributedTask.Pipelines.ObjectTemplating;
+using GitHub.Runner.Worker.Container.ContainerHooks;
 
 namespace GitHub.Runner.Worker
 {
